Systembugtj/issue73

+++ src/preload/file-config.ts
@@ -1,9 +1,8 @@
 import fs from "fs-extra";
 import path from "path";
-import type { PhotasaConfig, LoadCallback, PhotasaConfigResult } from "./types";
-
+import type { PhotasaConfig } from "./types";
+import { toFileName, shortenThumbnailName } from "../common";
 import * as R from "ramda";
-import { queryPhotasaConfigs } from "./query-config";
 import isVideo from "is-video";
 
 import { electronAPI } from "@electron-toolkit/preload";
@@ -11,15 +10,26 @@ const { ipcRenderer } = electronAPI;
 
 const PHOTASA_VERSION = "1.0";
 
-async function ensureConfig(photo: string, isFile: boolean): Promise<string> {
+export async function ensureConfig(
+    photo: string,
+    isFile: boolean,
+    isDelete = false,
+): Promise<string> {
     const dir = isFile ? path.dirname(photo) : photo;
     const configPath = path.join(dir, ".photasa.json");
-    await fs.ensureFile(configPath);
+    if (!isDelete) {
+        await fs.ensureFile(configPath);
+    }
+
     return configPath;
 }
 
-async function readConfig(photo: string, isFile: boolean): Promise<{ data: string; dir: string }> {
-    const dir = await ensureConfig(photo, isFile);
+export async function readConfig(
+    photo: string,
+    isFile: boolean,
+    isDelete = false,
+): Promise<{ data: string; dir: string }> {
+    const dir = await ensureConfig(photo, isFile, isDelete);
     const data = (await fs.readFile(dir, "utf-8")) ?? "{}";
     return {
         dir,
@@ -27,13 +37,21 @@ async function readConfig(photo: string, isFile: boolean): Promise<{ data: strin
     };
 }
 
-async function writeConfig(configPath: string, photoConfig: PhotasaConfig): Promise<void> {
+export async function writeConfig(
+    configPath: string,
+    photoConfig: PhotasaConfig,
+    isDelete = false,
+): Promise<void> {
+    // When delete a file, only write config if config is existed.
+    if (isDelete && !(await fs.exists(configPath))) {
+        return;
+    }
     photoConfig.lastModified = Date.now();
     const data = JSON.stringify(photoConfig, null, 4);
     await fs.writeFile(configPath, data, { encoding: "utf8", flag: "w" });
 }
 
-function fromJson(data: string): PhotasaConfig {
+export function fromJson(data: string): PhotasaConfig {
     try {
         return <PhotasaConfig>JSON.parse(data);
     } catch {
@@ -41,7 +59,7 @@ function fromJson(data: string): PhotasaConfig {
     }
 }
 
-function normalizeConfig(config: PhotasaConfig): PhotasaConfig {
+export function normalizeConfig(config: PhotasaConfig): PhotasaConfig {
     if (!config.photoList) {
         config.photoList = [];
     }
@@ -51,7 +69,7 @@ function normalizeConfig(config: PhotasaConfig): PhotasaConfig {
     return config;
 }
 
-const parseConfig = R.compose(normalizeConfig, fromJson);
+export const parseConfig = R.compose(normalizeConfig, fromJson);
 
 /**
  * Add photo to .photasa.json
@@ -72,13 +90,14 @@ export async function addToPhotoList(photoPath: string): Promise<void> {
 export async function removeFromPhotoList(
     photoPath: string,
 ): Promise<{ path: string; config: PhotasaConfig }> {
-    const meta = await readConfig(photoPath, true);
+    const meta = await readConfig(photoPath, true, true);
     const photasaConfig = parseConfig(meta.data);
-    const photoIndex = photasaConfig.photoList.findIndex((p) => p.path === photoPath);
+    const relative = toFileName(photoPath);
+    const photoIndex = photasaConfig.photoList.findIndex((p) => p.path === relative);
 
     if (photoIndex >= 0) {
         photasaConfig.photoList.splice(photoIndex, 1);
-        writeConfig(meta.dir, photasaConfig);
+        writeConfig(meta.dir, photasaConfig, true);
     }
     return {
         path: meta.dir,
@@ -108,24 +127,10 @@ export async function fixPhotasaConfig(folder: string): Promise<PhotasaConfig> {
     config.photoList.forEach((photo) => {
         photo.isVideo = isVideo(photo.path);
         photo.path = toFileName(photo.path);
-        photo.thumbnail = toThumbnailName(photo.thumbnail);
+        photo.thumbnail = shortenThumbnailName(photo.thumbnail);
     });
 
     writeConfig(meta.dir, config);
 
     return config;
-}
-
-export async function loadPhotasaConfigs(paths: string[], callback: LoadCallback): Promise<void> {
-    queryPhotasaConfigs(paths, (action, path) => {
-        callback(action, path);
-    });
-}
-
-export function toFileName(file: string): string {
-    return path.basename(file);
-}
-
-export function toThumbnailName(file: string): string {
-    return `.photasaoriginals/${path.basename(file)}`;
 }
